+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/data/ModRecipeProvider.kt
@@ -5,7 +5,9 @@ import net.minecraft.data.recipes.FinishedRecipe
 import net.minecraft.data.recipes.RecipeProvider
 import net.minecraft.world.item.crafting.Ingredient
 import site.siredvin.peripheralium.data.blocks.TweakedShapedRecipeBuilder
+import site.siredvin.peripheralium.data.blocks.TweakedShapelessRecipeBuilder
 import site.siredvin.peripheralium.data.blocks.TweakedUpgradeRecipeBuilder
+import site.siredvin.peripheralworks.common.setup.Blocks
 import site.siredvin.peripheralworks.common.setup.Items
 import site.siredvin.peripheralworks.xplat.ModRecipeIngredients
 import java.util.function.Consumer
@@ -29,5 +31,23 @@ class ModRecipeProvider(output: PackOutput) : RecipeProvider(output) {
             ingredients.netheriteIngot,
             Items.NETHERITE_PERIPHERALIUM_MODEM.get()
         ).save(consumer)
+
+        TweakedShapedRecipeBuilder.shaped(Blocks.PERIPHERAL_CASING.get().asItem())
+            .define('B', ingredients.peripheraliumBlock)
+            .define('C', ingredients.anyCoal)
+            .define('I', ingredients.ironIngot)
+            .pattern("ICI")
+            .pattern("CBC")
+            .pattern("ICI")
+            .save(consumer)
+
+        TweakedShapedRecipeBuilder.shaped(Blocks.UNIVERSAL_SCANNER.get().asItem())
+            .define('O', ingredients.observer)
+            .define('C', Ingredient.of(Blocks.PERIPHERAL_CASING.get().asItem()))
+            .define('P', ingredients.peripheralium)
+            .pattern("POP")
+            .pattern(" C ")
+            .pattern(" O ")
+            .save(consumer)
     }
 }