+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/PeripheralWorksCommonHooks.kt
@@ -1,14 +1,21 @@
 package site.siredvin.peripheralworks.xplat
 
 import dan200.computercraft.api.client.ComputerCraftAPIClient
-import dan200.computercraft.api.client.turtle.TurtleUpgradeModeller
 import dan200.computercraft.api.pocket.PocketUpgradeSerialiser
 import dan200.computercraft.api.turtle.TurtleUpgradeSerialiser
+import site.siredvin.peripheralium.computercraft.peripheral.owner.PocketPeripheralOwner
+import site.siredvin.peripheralium.computercraft.peripheral.owner.TurtlePeripheralOwner
+import site.siredvin.peripheralium.computercraft.pocket.BasePocketUpgrade
+import site.siredvin.peripheralium.computercraft.pocket.PeripheralPocketUpgrade
+import site.siredvin.peripheralium.computercraft.turtle.PeripheralTurtleUpgrade
 import site.siredvin.peripheralworks.PeripheralWorksClientCore
-import site.siredvin.peripheralworks.client.turtle.PeripheraliumHubModeller
+import site.siredvin.peripheralworks.client.turtle.ScaledItemModeller
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
+import site.siredvin.peripheralworks.common.setup.BlockEntityTypes
+import site.siredvin.peripheralworks.common.setup.Blocks
 import site.siredvin.peripheralworks.common.setup.Items
 import site.siredvin.peripheralworks.computercraft.peripherals.PeripheraliumHubPeripheral
+import site.siredvin.peripheralworks.computercraft.peripherals.UniversalScannerPeripheral
 import site.siredvin.peripheralworks.computercraft.pocket.PeripheraliumHubPocketUpgrade
 import site.siredvin.peripheralworks.computercraft.turtles.PeripheraliumHubTurtleUpgrade
 import java.util.function.Consumer
@@ -34,7 +41,7 @@ object PeripheralWorksCommonHooks {
             listOf(Consumer {
                 PeripheralWorksClientCore.registerHook {
                     ComputerCraftAPIClient.registerTurtleUpgradeModeller(
-                        it.get(), PeripheraliumHubModeller(0.5f)
+                        it.get(), ScaledItemModeller(0.5f)
                     )
                 }
             })
@@ -58,7 +65,29 @@ object PeripheralWorksCommonHooks {
             listOf(Consumer {
                 PeripheralWorksClientCore.registerHook {
                     ComputerCraftAPIClient.registerTurtleUpgradeModeller(
-                        it.get(), PeripheraliumHubModeller(0.5f)
+                        it.get(), ScaledItemModeller(0.5f)
+                    )
+                }
+            })
+        )
+
+        PeripheralWorksPlatform.registerTurtleUpgrade(
+            UniversalScannerPeripheral.UPGRADE_ID,
+            TurtleUpgradeSerialiser.simpleWithCustomItem { _, stack ->
+                return@simpleWithCustomItem PeripheralTurtleUpgrade.dynamic(
+                    stack.item,
+                    UniversalScannerPeripheral::of
+                ) { UniversalScannerPeripheral.UPGRADE_ID }
+            }, { dataProvider, serializer ->
+                dataProvider.simpleWithCustomItem(
+                    UniversalScannerPeripheral.UPGRADE_ID,
+                    serializer, Blocks.UNIVERSAL_SCANNER.get().asItem()
+                )
+            } ,
+            listOf(Consumer {
+                PeripheralWorksClientCore.registerHook {
+                    ComputerCraftAPIClient.registerTurtleUpgradeModeller(
+                        it.get(), ScaledItemModeller(0.5f)
                     )
                 }
             })
@@ -97,10 +126,27 @@ object PeripheralWorksCommonHooks {
                 serializer, Items.NETHERITE_PERIPHERALIUM_MODEM.get()
             )
         }
+
+        PeripheralWorksPlatform.registerPocketUpgrade(
+            UniversalScannerPeripheral.UPGRADE_ID,
+            PocketUpgradeSerialiser.simpleWithCustomItem { _, stack ->
+                return@simpleWithCustomItem PeripheralPocketUpgrade(
+                    UniversalScannerPeripheral.UPGRADE_ID,
+                    stack, UniversalScannerPeripheral::of
+                )
+            }
+        ) { dataProvider, serializer ->
+            dataProvider.simpleWithCustomItem(
+                UniversalScannerPeripheral.UPGRADE_ID,
+                serializer, Blocks.UNIVERSAL_SCANNER.get().asItem()
+            )
+        }
     }
 
     fun onRegister() {
+        BlockEntityTypes.doSomething()
         Items.doSomething()
+        Blocks.doSomething()
         registerTurtleUpgrades()
         registerPocketUpgrades()
     }