+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/fabric/FabricModRecipeIngredients.kt
@@ -3,6 +3,7 @@ package site.siredvin.peripheralworks.fabric
 import dan200.computercraft.shared.ModRegistry.Items
 import net.fabricmc.fabric.api.tag.convention.v1.ConventionalItemTags
 import net.minecraft.world.item.crafting.Ingredient
+import net.minecraft.world.level.block.Blocks
 import site.siredvin.peripheralworks.xplat.ModRecipeIngredients
 
 object FabricModRecipeIngredients: ModRecipeIngredients {
@@ -16,4 +17,12 @@ object FabricModRecipeIngredients: ModRecipeIngredients {
         get() = Ingredient.of(ConventionalItemTags.EMERALDS)
     override val diamond: Ingredient
         get() = Ingredient.of(ConventionalItemTags.DIAMONDS)
+    override val ironIngot: Ingredient
+        get() = Ingredient.of(ConventionalItemTags.IRON_INGOTS)
+    override val anyCoal: Ingredient
+        get() = Ingredient.of(ConventionalItemTags.COAL)
+    override val peripheraliumBlock: Ingredient
+        get() = Ingredient.of(site.siredvin.peripheralium.common.setup.Blocks.PERIPHERALIUM_BLOCK.get().asItem())
+    override val observer: Ingredient
+        get() = Ingredient.of(Blocks.OBSERVER)
 }