+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/ForgePeripheralWorks.kt
@@ -1,10 +1,13 @@
 package site.siredvin.peripheralworks
 
 import dan200.computercraft.api.ForgeComputerCraftAPI
-import dan200.computercraft.api.peripheral.IPeripheralProvider
+import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.item.Item
 import net.minecraft.world.level.block.Block
+import net.minecraft.world.level.block.entity.BlockEntityType
 import net.minecraftforge.common.util.LazyOptional
+import net.minecraftforge.event.CreativeModeTabEvent
+import net.minecraftforge.eventbus.api.SubscribeEvent
 import net.minecraftforge.fml.ModLoadingContext
 import net.minecraftforge.fml.common.Mod
 import net.minecraftforge.fml.config.ModConfig
@@ -13,6 +16,7 @@ import net.minecraftforge.registries.DeferredRegister
 import net.minecraftforge.registries.ForgeRegistries
 import net.minecraftforge.registries.NewRegistryEvent
 import site.siredvin.peripheralium.ForgePeripheralium
+import site.siredvin.peripheralium.api.peripheral.IPeripheralProvider
 import site.siredvin.peripheralworks.common.Registries
 import site.siredvin.peripheralworks.common.configuration.ConfigHolder
 import site.siredvin.peripheralworks.computercraft.ComputerCraftProxy
@@ -26,12 +30,15 @@ import thedarkcolour.kotlinforforge.forge.MOD_CONTEXT
 
 
 @Mod(PeripheralWorksCore.MOD_ID)
+@Mod.EventBusSubscriber(modid = PeripheralWorksCore.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 object ForgePeripheralWorks {
 
     val blocksRegistry: DeferredRegister<Block> =
         DeferredRegister.create(ForgeRegistries.BLOCKS, PeripheralWorksCore.MOD_ID)
     val itemsRegistry: DeferredRegister<Item> =
         DeferredRegister.create(ForgeRegistries.ITEMS, PeripheralWorksCore.MOD_ID)
+    val blockEntityTypesRegistry: DeferredRegister<BlockEntityType<*>> =
+        DeferredRegister.create(ForgeRegistries.BLOCK_ENTITY_TYPES, PeripheralWorksCore.MOD_ID)
 
     init {
         ForgePeripheralium.sayHi()
@@ -49,6 +56,7 @@ object ForgePeripheralWorks {
         PeripheralWorksCommonHooks.onRegister()
         blocksRegistry.register(eventBus)
         itemsRegistry.register(eventBus)
+        blockEntityTypesRegistry.register(eventBus)
         Registries.TURTLE_SERIALIZER.register(eventBus)
         Registries.POCKET_SERIALIZER.register(eventBus)
     }
@@ -59,15 +67,26 @@ object ForgePeripheralWorks {
         Platform.maybeLoadIntegration("easy_villagers").ifPresent { (it as Runnable).run() }
         Platform.maybeLoadIntegration("toms_storage").ifPresent { (it as Runnable).run() }
         // Register peripheral provider
-        ForgeComputerCraftAPI.registerPeripheralProvider(IPeripheralProvider { world, pos, side ->
+        ForgeComputerCraftAPI.registerPeripheralProvider { world, pos, side ->
+            val entity = world.getBlockEntity(pos)
+            if (entity is IPeripheralProvider<*>) {
+                val foundPeripheral = entity.getPeripheral(side)
+                if (foundPeripheral != null)
+                    return@registerPeripheralProvider LazyOptional.of { foundPeripheral }
+            }
             val supplier = ComputerCraftProxy.lazyPeripheralProvider(world, pos, side)
-                ?: return@IPeripheralProvider LazyOptional.empty()
-            return@IPeripheralProvider LazyOptional.of { supplier.get() }
-        })
+                ?: return@registerPeripheralProvider LazyOptional.empty()
+            return@registerPeripheralProvider LazyOptional.of { supplier.get() }
+        }
     }
 
     fun registrySetup(event: NewRegistryEvent) {
         Platform.maybeLoadIntegration("integrateddynamics").ifPresent { (it as Runnable).run() }
         Platform.maybeLoadIntegration("naturescompass").ifPresent { (it as Runnable).run() }
+    }
+
+    @SubscribeEvent
+    fun registerCreativeTab(event: CreativeModeTabEvent.Register) {
+        event.registerCreativeModeTab(ResourceLocation(PeripheralWorksCore.MOD_ID, "tab"), PeripheralWorksCore::configureCreativeTab)
     }
 }