+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/forge/ForgeModRecipeIngredients.kt
@@ -2,6 +2,7 @@ package site.siredvin.peripheralworks.forge
 
 import dan200.computercraft.shared.ModRegistry.Items
 import net.minecraft.world.item.crafting.Ingredient
+import net.minecraft.world.level.block.Blocks
 import net.minecraftforge.common.Tags
 import site.siredvin.peripheralworks.xplat.ModRecipeIngredients
 
@@ -16,4 +17,12 @@ object ForgeModRecipeIngredients: ModRecipeIngredients {
         get() = Ingredient.of(Tags.Items.GEMS_EMERALD)
     override val diamond: Ingredient
         get() = Ingredient.of(Tags.Items.GEMS_DIAMOND)
+    override val ironIngot: Ingredient
+        get() = Ingredient.of(Tags.Items.INGOTS_IRON)
+    override val anyCoal: Ingredient
+        get() = Ingredient.of(net.minecraft.world.item.Items.COAL)
+    override val peripheraliumBlock: Ingredient
+        get() = Ingredient.of(site.siredvin.peripheralium.common.setup.Blocks.PERIPHERALIUM_BLOCK.get())
+    override val observer: Ingredient
+        get() = Ingredient.of(Blocks.OBSERVER)
 }