+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/forge/ForgePeripheralWorksPlatform.kt
@@ -7,15 +7,20 @@ import dan200.computercraft.api.turtle.ITurtleUpgrade
 import dan200.computercraft.api.turtle.TurtleUpgradeDataProvider
 import dan200.computercraft.api.turtle.TurtleUpgradeSerialiser
 import dan200.computercraft.api.upgrades.UpgradeDataProvider
+import net.minecraft.core.BlockPos
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.item.Item
 import net.minecraft.world.level.block.Block
+import net.minecraft.world.level.block.entity.BlockEntity
+import net.minecraft.world.level.block.entity.BlockEntityType
+import net.minecraft.world.level.block.state.BlockState
 import site.siredvin.peripheralworks.ForgePeripheralWorks
 import site.siredvin.peripheralworks.PeripheralWorksCore
 import site.siredvin.peripheralworks.common.Registries
 import site.siredvin.peripheralworks.data.ModPocketUpgradeDataProvider
 import site.siredvin.peripheralworks.data.ModTurtleUpgradeDataProvider
 import site.siredvin.peripheralworks.xplat.PeripheralWorksPlatform
+import java.util.Set
 import java.util.function.BiFunction
 import java.util.function.Consumer
 import java.util.function.Supplier
@@ -26,10 +31,28 @@ object ForgePeripheralWorksPlatform: PeripheralWorksPlatform {
     }
 
     override fun <T : Block> registerBlock(key: ResourceLocation, block: Supplier<T>, itemFactory: (T) -> Item): Supplier<T> {
-        PeripheralWorksCore.LOGGER.warn("Register block", )
+        PeripheralWorksCore.LOGGER.warn("Register block")
         val blockRegister = ForgePeripheralWorks.blocksRegistry.register(key.path, block)
         ForgePeripheralWorks.itemsRegistry.register(key.path) { itemFactory(blockRegister.get()) }
         return blockRegister
+    }
+
+    override fun <V : BlockEntity, T : BlockEntityType<V>> registerBlockEntity(
+        key: ResourceLocation,
+        blockEntityTypeSup: Supplier<T>
+    ): Supplier<T> {
+        return ForgePeripheralWorks.blockEntityTypesRegistry.register(key.path, blockEntityTypeSup)
+    }
+
+    override fun <T : BlockEntity> createBlockEntityType(
+        factory: BiFunction<BlockPos, BlockState, T>,
+        block: Block
+    ): BlockEntityType<T> {
+        return BlockEntityType.Builder.of({ t: BlockPos?, u: BlockState? ->
+            factory.apply(
+                t!!, u!!
+            )
+        }, block).build(null)
     }
 
     override fun <V : ITurtleUpgrade> registerTurtleUpgrade(